2.1.0 bugfixes, $v update earlydev
A lot of things needed to be asynced and a lot of things needed to be awaited. A lot of things also needed to be selfed. There are still some obvious bugs but they're in commands I'm not working on yet. $v update works, but not well, I wrote it before I understood how .jsons actually could be used.